Move common config interfaces to core
